+++ src/main/java/com/pauluno/finledger/domain/model/AccountStatus.java
@@ -1,0 +1,6 @@
+package com.pauluno.finledger.domain.model;
+
+public enum AccountStatus {
+    OPEN,
+    CLOSED
+}